JORSON: Be able to add buildings to the grid.

+++ Helpers/P5/P5App.Environment.cs
@@ -34,7 +34,7 @@ namespace ExoKomodo.Helpers.P5
             "frameCount"
         );
         
-        public uint Height => _jsRuntime.Invoke<uint>(
+        public float Height => _jsRuntime.Invoke<uint>(
             _p5GetValue,
             "height"
         );
